connect modules Trace and Inference
Issue:
MH isn't correct on programs with conditioning, grassModel
for example. Investigate.
