+++ monad-bayes.cabal
@@ -18,6 +18,7 @@ library
   exposed-modules:     Primitive, Base, Dist, Sampler, Prior,
                        Rejection, Weighted, Empirical, Particle,
                        Inference, Metrics
+  other-modules:       Trace.Debug
   build-depends:       base >= 4.7 && < 5
                      , containers
                      , random
